+++ src/Worker.java
@@ -1,4 +1,4 @@
 public class Worker extends Employee {
 
-    protected double salary;
+    protected int salary;
 }
